:zap:Ignore yarn engines
Yarn will die if it is using an incompatible node version from the
version in specified `package.json`. This is useful for applications
that need to run (like bigtest server), but is not super necessary
when all that you want to do is lint.

This sets the `--ignore-engines` flag when running yarn so whatever
version is on the docker container will be used.